Update patch set 2
Patch Set 2: Code-Review+2

Thanks. Please make sure in the future to explain the real motivation for a change, either in the CL description or by citing a relevant bug. (In this case, #11211 is closed as won't fix, and I don't think this CL fixes it anyway.)

Your follow-up explanation makes sense to me. In your case you end up with a .git file instead of a .git directory. I'm happy to use the git rev-parse to handle this. And now I know about git worktrees. :-)

Patch-set: 2
Label: Code-Review=+2