Update patch set 2
Patch Set 2:

(1 comment)

This seems off still. The new bool is essentially overriding the information in some combination of level, src, srcE, dst, dstE. Why is that information not right, and can it be corrected? In particular, for things that escape to the heap, I expected to see escwalk called with dst == sinkNode. And if those are incorrect, the short-circuit return at the top of escwalk (now escwalkBody) based only on level comparison may do the wrong thing and bypass a body that needs to be reexecuted due to the changed bool. If we can get the other state corrected then that's not a concern.

Patch-set: 2
Label: Code-Review=0