Update patch set 2
Patch Set 2:

> This seems off still. The new bool is essentially overriding the
 > information in some combination of level, src, srcE, dst, dstE. Why
 > is that information not right, and can it be corrected?

The canonical approach to this is to "track" things of importance so that they become the source for "escflood", but the closure (I think) escapes because of loop level issues and somehow ends up not tracked.  I'll look into that some more.  Adding escHeap makes this instead semi-iterative.

A different semi-iteraative approach might be to use a slice to record e.dsts and add newly exposed OCLOSURE nodes to it as they are discovered, and then there are just a few more flood steps.

The overriding with the "escaped" parameter is relatively lightweight -- it is at least only more conservative, yet does not fire incorrectly in our tests, so that part doesn't bother me too much.  The information that it adds to "leaks" is reasonable -- "this is heap-exposed".

The failure to take it into account in the early-out code does bother me; it's not a problem from flood to flood because of the walkgen guard, but there might be a problem in a single flood.

> In
 > particular, for things that escape to the heap, I expected to see
 > escwalk called with dst == sinkNode. And if those are incorrect,
 > the short-circuit return at the top of escwalk (now escwalkBody)
 > based only on level comparison may do the wrong thing and bypass a
 > body that needs to be reexecuted due to the changed bool. If we can
 > get the other state corrected then that's not a concern.

Patch-set: 2