Update patch set 1
Patch Set 1:

> I'm ok with this change, but are there really stack barriers on the
 > g0 stacks? I thought they'd just be on the user stack (curg), in
 > which case the locking might have been overconservative but ok
 > before.

There are not stack barriers on the g0 stacks, but the locking was *not* okay before.

In all three cases we can unwind from a user g (mp.curg in the cgo case, and libcallg in the libcall case), but currently the cgo and libcall cases can happen without taking the stack barrier lock.

(It's possible that with this change the locking becomes overconservative because there may be implications like "if a G is in a libcall, nothing can be modifying its stack barriers", but that just guarantees that we'll be able to grab the stack barrier lock.)

Patch-set: 1