Update patch set 2
Patch Set 2:

No joy in mudville:

1) tried to infect the captured variables of a closure with the loop level of the closure (passing in a one-shot override for the loop level).  Didn't work.

2) added escaped closures on to the list of top-level things to (re) walk.  Didn't work.

3) "fixed" this patch to include a "have I been here before with 'escaped' true?" to ensure necessary rewalking -- still worked, but caused a couple of failures in the escape regression tests.

Two things to try further:
a) Pass the loop depth (from closures)  all the way down.
b) limit the escape-flag set to only occur from escaped closures.

I think this is a hack, I think the "correct" approach might follow what I seem to be doing in the "explain" code where I build a more full-featured graph, and include all the relevant information on the edges of that graph instead of spattering them around the escape analysis code -- but that would require a rewrite of escape analysis (I am sure this is not good news to Chris).

Patch-set: 2