Update patch set 3
Patch Set 3: Run-TryBot+1

This is the improved hack -- still use the 'escaped' parameter, but with a rewalk check, and also limited only to starting at closures that happen to escape.

I think we'd be better off if we modeled closures more like how they are implemented, but again, that's a large change for 1.6; this is conservative with respect to escapes.

Patch-set: 3
Label: Run-TryBot=+1