Update patch set 6
Patch Set 6:

Hi,

the AGM method is based on the fact that if z is big enough, then

log(z) ≈ π / (2 * AGM(1, 4/z))

AGM(a,b) is the arithmetic–geometric mean of a and b, and can be
computed by iterating a = (a+b)/2, b = sqrt(a*b) until a == b.
The last thing you need is a fast way to compute π to arbitrary
precision, but that's not hard (Gauss–Legendre is fast enough).

A reference is: R. P. Brent, Multiple-precision zero-finding methods 
and the complexity of elementary function evaluation, in Analytic 
Computational Complexity, Academic Press, New York, 1975.

It's an old article, but the author has retyped it and it's now on
arXiv, here: http://arxiv.org/pdf/1004.3412.pdf.

Log is at section 9. 

My implementation is at github.com/ALTree/floatutils. I don't guarantee
it is correct, and it's a work-in-progress (and not much "in-progress" 
at the moment), but it seems to work.

Patch-set: 6
Reviewer: Gerrit User 5385 <5385@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0