runtime: two experiments in microoptimizing scanobject
DO NOT REVIEW.

This implements two experiments in improving the microarchitectural
behavior of scanobject. Unfortunately, neither help.

The first experiment is to reduce the number of unpredictable branches
in scanobject. scanobject is 13% of total branches in the garbage
benchmark and most of them are highly unpredictable. This experiment
eliminates the branches for deciding on pointer/scalar and for
detecting whether a pointer is non-nil, points in to the heap, and
doesn't point to the same object. After this, by far the most
unpredictable branch (not only in scanobject, but in the whole
program) is the loop over the length of the object itself. Overall
this has no discernable effect on garbage.

The second experiment is to scan batches of objects with the intent of
hiding memory latency. This introduces a new function, scanobjects,
that has the same signature as scanobject, but scans 16 objects
simultaneously. When it reaches the end of an object, it fetches a new
object in to that slot. In order to integrate with the exit conditions
in the callers, it stops refilling its pipeline after scanning 1MB of
data. This experiment slows down garbage by about 2X. This appears to
be because of the increased complexity of the code and, in particular,
the much larger number of unpredictable branches.

Change-Id: I6b9b1433d9f997ada63a3635fca64b149c41a440