Update patch set 1
Patch Set 1:

R=go1.7

I'm going to leave this for Go 1.7. It's been like this for so long we'll survive another release. The variable in MemStats is actually reporting whether we're so early in the runtime startup that GC hasn't been turned on yet. Even when GOGC=off, if EnableGC=true, then it's still enabled in the sense that you could call SetGCPercent and it would run. When EnableGC=false it means it is very very early. 

I think probably the right fix for this particular line is to delete it from the profile.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0