Update patch set 9
Patch Set 9:

R=go1.7

Sorry, it's too late. But this whole package is easily forkable on github or in your own project wherever. It's not a core piece of Go. Don't let this bug block you for anything.

When somebody has time to review we can fix it in the Go 1.7 dev cycle starting in Feb.

Patch-set: 9