archive/zip: clarify guarantees and expectations in RegisterCompressor and RegisterDecompressor
The example provided in archive/zip for Writer.RegisterCompressor relies
on expectations that are not clearly documentated in the RegisterCompressor
method. Thus, we explicitly say that Close will be called between each invocation
of comp.

Also, we document the expectation that the dcomp provided to
Reader.RegisterDecompressor must support being called concurrently if the user
plans on calling Open concurrently.

Updates #8359

Change-Id: Ia407b581dd124185f165c25f5701018a8ce4357a