net/http: copy request headers when following redirects
When headers are set on the request, caller expects those headers to be
passed across the redirected requests.

The Cookie header is excluded as the it would conflict with the cookie
jar and leak cookies across domains.

When a POST is redirected and transformed into a GET, the Content-Length
and Content-Type headers are removed as they are no longer relevent.

Fixes #4800

Change-Id: I26a2fb2ddebf1e65098dcb5fb2f1ac73b80efd6b