Update patch set 2
Patch Set 2:

(3 comments)

Well spotted.

In the documentation you write that "except for the first file...". So is the example incorrect?

I would rather that the API didn't provide any such guarantees, and we instead fix the example and document that the Compressor must also be concurrency safe. That would also allow bigger flexibility in the future, if we should decide to change the behavior.

Does that make sense?

Patch-set: 2
Label: Code-Review=0