Update patch set 9
Patch Set 9:

R=go1.7

(Sorry, that has to be the last comment to remove it from our dashboard)

The other thing I forgot to mention is that this is the sort of change that's likely to break existing tests because people use DumpRequest in tests often. We already spent a month+ cleaning up Google tests to make Go 1.6 the default inside Google, which is a requirement to doing a public release (that we're confident with the results ourselves).  Given that 1.6beta2 went out yesterday, we can't be changing behavior at this point.

Patch-set: 9