Update patch set 1
Patch Set 1:

> > the test later uses it in a string compare with an address that
 > is
 > > equivalent, but without the embedding, so the test fails.
 > 
 > Are you sure about test still failing? If you skip tests for
 > interfaces that are down, you won't have to check addresses with
 > "ipv4 embedded address". No?
 > 
 > Alex

 > > the test later uses it in a string compare with an address that
 > is
 > > equivalent, but without the embedding, so the test fails.
 > 
 > Are you sure about test still failing? If you skip tests for
 > interfaces that are down, you won't have to check addresses with
 > "ipv4 embedded address". No?
 > 
 > Alex

You're absolutely right. Skipping the down interfaces alone stops all the failures. However, in my testing, re-parsing those ipv4 embedded addresses alone fixed some of the failures, so I left it in as a partial improvement. I would be happy to remove it if you prefer (and as a side effect, not drag in regexp).

Patch-set: 1
Label: Code-Review=0