Update patch set 1
Patch Set 1:

> > To be honest I don't understand why this patch fixes the actual
 > > problem.  I don't see why this would affect any test that doesn't
 > > use cgo.  Do you understand what is happening?
 > 
 > not sure. for now what i know is that we can observe ss_flags==0 or
 > ss_flags==1 or ss_flags=4 in minit and kernel machine-dependent
 > stuff on BSD variants tends to use SS_ONSTACK for notification and
 > SS_DISABLE for configuration.

I think I may see it.  On NetBSD, and presumably on DragonFly, a newly created thread inherits the signal stack of the creating thread.  This does not happen on GNU/Linux and probably doesn't happen on other systems either.

The key change in your patch is that you are changing minit to always call signalstack for a newly created thread.

Patch-set: 1