Update patch set 4
Patch Set 4:

> (3 comments)

I am happy to test out the subset suggestion, but we seem to be seeing artifacts of how wireless interfaces are brought up and down, and subsequently reported by netsh. It might be too much to expect consistency from what various tools report as the address of a down interface. And in the end, if an interface falls in the woods, and there is no-one around to hear it, does its address matter?

Patch-set: 4
Reviewer: Gerrit User 12012 <12012@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0