runtime: test SS_ONSTACK, not SS_DISABLE, in sigtrampgo
The POSIX docs for sigaltstack are somewhat vague, but it seems fairly
clear that sigaltstack will return SS_ONSTACK when actually running on
the signal stack.  Since sigtrampgo is trying to determine whether we
are currently running on the signal stack, test SS_ONSTACK rather than
SS_DISABLE.

The test for this is in misc/cgo/testcarchive, but it is not currently
enabled for any system where it matters.

Change-Id: I6ab12eac1c69ceb3eb5594683d6505f0d6f2f3b0