Update patch set 4
Patch Set 4:

I know this was a while ago, but was there any discussion about the design decisions here?

Specifically, the decision seems to have been made unremarked that VEX-encoded instructions in the Go assembler will look like

    VTHING X1, X2, X3

but in fact emit

    VTHING ymm1, ymm2, ymm3

instead of

    VTHING xmm1, xmm2, xmm3.

This is surprising because historically X1 is the Go syntax for xmm1 not ymm1, and there are real albeit small and subtle differences between VTHING ymm and VTHING xmm. The syntax of the assembler is already more different from the manuals than it probably should be, and the decision made here seems like it makes that problem worse. If someone actually did want the xmm1 version it's unavailable because the obvious spelling is taken.

After thinking about this for a few hours I am inclined to introduce actual Y registers to the assembler so that both of these are valid:

    VTHING X1, X2, X3
    VTHING Y1, Y2, Y3

The alternative would be to put the decision in the opcode:

    VTHING.X X1, X2, X2
    VTHING.Y X1, X2, X3

but on balance it seems like following the manuals more closely will be more useful and less surprising to users writing assembly.

We need to figure this out soon so that we don't ship unfortunate assembly semantics in a release and get stuck with them forever.

I filed golang.org/issue/14068 to figure out what we want to do about this for Go 1.6.

Patch-set: 4
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0