Update patch set 1
Patch Set 1:

I think this comment belongs in the runtime/pprof package.

There's nothing in the pprof package docs that references SIGPROF. The signal is an undiscussed implementation detail of profiling, so it seems odd that the documentation for why profiling isn't working is in the signal package.

Patch-set: 1
Label: Code-Review=0