Update patch set 1
Patch Set 1:

I'm surprised less about the semantics than changing them so close to a release.

But on the semantics, I looked a bit more. There's no reverse conversion,right? So if you have a database with a time-typed column, you have to pass in a time.Time, and until now you also had to read back a time.Time. This CL would let you read back a string instead but you still can't pass in a string? Or is it that the individual drivers all parse strings behind the scenes?

Patch-set: 1