[dev.garbage] runtime: add stackfreelist
The freelist for normal objects and the freelist
for stacks share the same underlying type definitions
but are operated on by different code sequences. This
overloading complicates the use of bit vectors
for allocation of normal objects. This change refactors
the use of the stackfreelist out from the use of freelist.

Change-Id: I5b155b5b8a1fcd8e24c12ee1eb0800ad9b6b4fa0