Update patch set 3
Patch Set 3:

Here's an example that is missed.   I'm going to see if I can figure out what's (not) happening here today, if not, your turn this evening.  I'm not in a tearing rush because I think bounds check elimination will require a lot of reviewing.

Note that I am running CSE early for BCE, before SliceLen is lowered, so you might not see this.  The non-commoned operation is SliceLen(b):

package foo

func h(a, b []int) []int {
	if len(a) > len(b) {  // SliceLen(b) here
		return h(b, a)
	}
	n := len(a)
	c := make([]int, n)  // Analysis doesn't see this.
	if len(c) < len(a) { // Necessary to establish acceptable len(a)
		panic("Impossible") // Does not return, necessary
	}
	for i := 0; i < n; i++ {
		aa := a[i]     // ERROR "removed bounds check"
		bb := b[i]     // ERROR "removed bounds check" // SliceLen(b) here
		c[i] = aa + bb // ERROR "removed bounds check"
	}
	return c
}

Patch-set: 3