Update patch set 1
Patch Set 1:

My test case was 

package foo

func h(a, b []int) []int {
	if len(a) > len(b) {
		return h(b, a)
	}
	n := len(a)
	c := make([]int, n)  // Analysis doesn't see this.
	if len(c) < len(a) { // Necessary to establish acceptable len(a)
		panic("Impossible") // Does not return, necessary
	}
	for i := 0; i < n; i++ {
		aa := a[i]     // ERROR "removed bounds check"
		bb := b[i]     // ERROR "removed bounds check"
		c[i] = aa + bb // ERROR "removed bounds check"
	}
	return c
}

I realized just now (in the shower, of course) that this doesn't break transitive CSE; the problem was because a and b are args.

I think the real problem was that partition made finer distinctions than sort, therefore, sort would place things in order SliceLen(b), SliceLen(a), SliceLen(a), SliceLen(b) and then partition would kick the two SliceLen(b) out into separate partitions.  But since Aux is not comparable we don't use it in sort, so we can't use it in partition.

Patch-set: 1