Update patch set 5
Patch Set 5:

I tried allocating the histograms on the stack as proposed by Joe Tsai. Contrary to my expectations it didn't make any positive difference. In fact it was a tiny bit slower. It might be some difference in copy vs. zeroing that makes the difference.

However, it did point out some unneeded zeroing that was done in "reset". Various slices/structs are zeroed, even though they are assumed to be dirty when the encoder is called.

I removed this along with some minor changes. This should address all issues you asked me to look at so far.

Patch-set: 5