Update patch set 5
Patch Set 5:

Another general comment is that, according to the CL description itself, this does four things, for some quantum of improvement (say roughly +10%).

Some of these four things are relatively simple, others are relatively complicated. They also look, at first glance, like orthogonal things: e.g. "removing allocations" sounds like it is separable from "accumulate 48 bits instead of 24". It would be nice to know the breakdown of how each of those four things contribute to the +10%.

If it turns out that things A, B and D are all relatively simple code changes and gets you +9%, and C is very complicated and gets you +1%, I would recommend rejecting C even if that +1% was statistically significant. Raw MB/s throughput is not the only criteria in which the Go standard library code is judged. As this CL is packaged now, though, we either have to accept all four or reject all four.

Accordingly, my recommendation would be to break this CL up into a smaller ones, even if it is already listed as "part 1 of many". As always, smaller yet still self-contained CLs are easier to review, and easier to manage (e.g. roll back if necessary).

Patch-set: 5