Update patch set 5
Patch Set 5:

(1 comment)

> this does four things, for some quantum of improvement 

I wouldn't call 10%, as well as a removal of all allocations a quantum improvement. As other parts of the code will be sped up, this improvement will become a much larger.

>Accordingly, my recommendation would be to break this CL up into a smaller ones, even if it is already listed as "part 1 of many". 

The granularity of the changes can be discussed, and this was proposed privately by Ross Cox.

Patch-set: 5