Update patch set 5
Patch Set 5:

I'm not saying that we need a test that 100% of the time sets up the AX and DX registers to fail. I'm saying that it's not hard to write a test that an asm function and a pure Go function behave identically on a variety of inputs, whether randomly generated or explicitly designed. For example, see https://go.googlesource.com/exp/+/master/shiny/driver/internal/swizzle/swizzle_test.go

Even if such a test might not have picked up the AX/DX bug, or some other bug we don't know about, with 100% certainty, I'm pretty sure there'd be a significantly non-zero chance that it could have. On the other hand, if you don't explicitly test the new matchLenSSE4 function, you have a much smaller, if not practically zero, chance of picking up a bug like this early. If that bug was only discovered later, it's much harder to debug "these thousands of lines of code has a bug somewhere" than it is to debug "this hundred line patch has a bug somewhere", especially harder if you're debugging asm code, as your recent golang-nuts post illustrates. Maybe there are other bugs lurking that we don't know about yet, which we could shake out from a simple unit test. That seems especially worthwhile if such a test is not that hard to write.

You could argue that there are existing tests that the output is 'well compressed', yet they did not catch this AX/DX bug, even though the effect of that bug means that your compression rate falls. I therefore think that it's worthwhile to write some more tests, whether end-to-end "compress this input" tests or more focused tests of "calculate the longest match" functions, or both.

For example, I recently wrote https://github.com/golang/snappy/commit/d1d908a252c22fd7afd36190d5cffb144aa8f777 that fixed a bug where some output was not as well compressed as it should be (and the MB/s throughput benchmark was therefore artificially high). But it did require writing a new test: one in which the first half of the input was very incompressible and the second half is very compressible. That was for snappy, and this is flate, and while the TestDeflateInflateString you pointed to doesn't try any inputs shaped like that (only two, very compressible text files), I'm not saying that the flate package necessarily needs exactly that same test. Instead, that new snappy test was written after thinking about when a recent change would and wouldn't take effect. Here, if it's too hard to think of a TestDeflateInflateString style round-trip test that similarly exercises the corner cases of the optimizations you are making, I think you can still make some more focused unit tests, e.g. for matchLenSSE4.

Sorry for the long rant. I do think that you have made some genuine performance improvements, I appreciate the effort you've made to upstream them, and I'd love to see them land in the standard library. I'm just wary of inadvertently introducing bugs while optimizing, and it seems to me that an easy way to make that far less likely is to add new tests along with new code, especially if it's clever code, as optimized code tends to be.

As for the granularity of the changes, Russ is more than qualified to advise, and you've said that you've already been discussing it with him. I'll leave that decision for the two of you, after, of course, he's done with the imminent Go 1.6 release.

P.S. Sorry, I should have been clearer. By "quantum of improvement", I didn't mean "a small amount of improvement". Instead, I meant "some numeric value of improvement", and my general point was that it could be broken down and have separate values accounted to code changes A, B, C and D, so that a cost-benefit discussion could apply to each part. In any case, Russ is perfectly capable of continuing that conversation if he wants to.

Patch-set: 5
Label: Code-Review=0