Update patch set 6
Patch Set 6:

I have uploaded a new patch, which updates the assembler, and reverts a change which didn't provide a consistent speedup.

Furthermore I have written a regression test, which checks for changes in the encoding patters of the bit encoders/writer. This can be used to later check for unintended encoding changes.

You can see it as a PR here: https://github.com/klauspost/compress/pull/40 - I will be adding it to this CL, probably tomorrow. If you see it before that and have comments, feel free to leave it as PR comments.

Patch-set: 6
Label: Code-Review=0